Add functionality for socket communication.
The Message module contains read and write functions that take care of
the wire protocol, which sends a 4 byte length header before every message.
They take a bytes buffer and don't care about its contents.

The Vyconf_client module has high level functions for interfacing with vyconfd.

Functions for creating a socket and creating a server loop are in Startup.

The rest is, for now at least, right in vyconfd.

Notes:
Message.read/write probably should be wrapped in Lwt_io.atomic
We need to find out if it's safe to reuse Pbrt.Encoder.t or we really
should create it every time.