Render valueless nodes correctly.
The canonical format for valueless nodes is "foo;".
There are two cases: when it's known from the reference tree that
a node is valueless, we should render it the canonical way.

However, a valueless node is not the same as a normal node with no children,
semantically, so nodes that are note valueless should be rendered as "foo {}".

If the reference tree is not present, resort to rendering all nodes
without values as if they were normal nodes with no children
rather than as if they were valueless.

The parser is unaware of the distinction, so there are no implications for it.