T6718: use correct lib name in absence of libvyatta-cfg-dev
An artifact of development was the use of lib name libvyatta-cfg.so for
dlopen instead of libyatta-cfg.so.1. The former exists when the
libvyatta-cfg-dev package is installed, as it is when vyatta-cfg is
built locally. As the dev package is not needed nor used by default
builds, correct to the latter lib name, as installed by libvyatta-cfg.